+++ Assets/Scripts/Player/ISavedPlayerData.cs
@@ -5,14 +5,14 @@ public interface ISavedPlayerData
     public IReadOnlyCollection<IPlayerPurchase> OneTimePurchases {  get; }
     public float HealthCoefficient { get; }
     public float DamageCoefficient { get; }
-    public int CountCoins { get; }
+    public long CountCoins { get; }
     public int AchievedLevelIndex { get; }
 
     public void SetHealthCoefficient(float healthCoefficient);
 
     public void SetDamageCoefficient(float damageCoefficient);
 
-    public void SetCountCoins(int countCoins);
+    public void SetCountCoins(long countCoins);
 
     public void SetAchievedLevelIndex(int achievedLevelIndex);
 }
