+++ Assets/Scripts/SDK/Saves/PlayerProgress.cs
@@ -4,7 +4,7 @@
     {
         public float HealthCoefficient = 1f;
         public float DamageCoefficient = 1f;
-        public int CountCoins = 1000;
+        public long CountCoins = 1000;
         public int AchievedLevelIndex = 0;
     }
 }
